Thread-lock the loan-out operation
The default Bottle dev web server is single-thread, so we won't run into
the problem of 2 users simultaneously clicking on the loan button.  Other
servers are multithreaded, and there's a risk that the time it takes us
to look through the loans introduces a window of time when another user
might click on the same loan button and cause another loan request to be
initiated before the 1st finishes.  So, lock this block of code.