Refactor multiple things for better WSGI experience
This introduces multiple changes.

One is that `run-server` is rewritten to use the Python mod_wsgi package and always runs the local server using `adapter.wsgi` and Apache mod_wsgi, instead of the built-in Bottle server.  The previous `Server()` object is gone from `server.py`.  The approach of always using `adapter.wsgi` should be better for testing and developemnt.

A problem with the wsgi scheme of things is that the base path is set by the apache configuration file, and not by the application code.  This created the problem that settings like base_url and the template path in `settings.ini` were ignored.  I was afraid this would be source of perpetual confusion and errors, so I worked out how to get the runtime base url from the server itself (eliminating some code and one `settings.ini` variable) and made the dibs server code set the template path itself.  Basically, I was trying to remove things from settings.ini that might be overriden by an Apache server configuration.

The previous `-@` flag is gone from `run-server` in favor of `-v` for `--verbose` and an different `-d` for `--debug`.  It now outputs always to the terminal if given either `-v` or `-d`.  The info getting logged is higher due to more things coming from the Apache wsgi server, although it's also needlessly detailed because I couldn't get certain log format configuration variables to work. (Even ended up posting on Stack Overflow to ask about it.)  There is both a `-v` and a `-d` option to `run-server` because `-d` debug mode in mod_wsgi **turns off auto-reloading upon source code changes** and there's nothing we can do about it.  (Why does it do that?  Beats me.)  So, basically the scheme now is:

1. if you want to get verbose logging and get auto-reloading, use `-v`
2. if you want to use a debugger like pdb use `-d`, but watch out that it won't reload files automatically