Complete overhaul of run-server
The annoying behavior of `mod_wsgi` with respect to debug mode
(particularly its disabling of auto-reload if debug mode is enabled):
frustrated me no end, and so finally I rewrote this to go back to
using Bottle's default server for debug mode.  This required a change
to get rid of the previous `DEBUG` config variable in "settings.ini",
in favor of a new variable, `RUN_MODE`.  This also meant a change to
the command-line flags for `run-server`, and a nearly complete
refactoring of the code.