Remove @head_method_ignored and handle it differently
It's not entirely correct to ignore HEAD methods, so I removed the
hacky @head_method_ignored annotator.  However, it turns out that we
can't properly handle HEAD anyway: for reasons that I haven't been
able to figure out, request.session lacks the user entry in HEAD
method invocations, which means that we can't check if the user has
been authenticated, which means that we have no choice except to
return nothing for HEAD in some cases anyway.  The situation where it
matters is in the @authenticated annotator, so that's where I moved
the check.