Be more careful about deletions while iterating
It seems like deleting an item from the result of a Thing.select()
call (e.g., Recent.select()) may be problematic.  I'm not 100% sure
but it seemed to produce errors in my testing, and that might make
sense depending on how the results of Recent.select() are produced.
To avoid the risk, let's create a list of the results first and then
iterate over the list we created.