Make update_item() much more efficient
I don't know why I originally wrote it to always try to get the item
from TIND, but it doesn't make sense to do it for /update/edit operations.
By skipping it for edit, we can do that operation much more quickly,
and we don't lose any error detection. Win-win.