Rewrite URLs to work around Apache limitation
The scheme to route all IIIF requests through the DIBS server worked
when using Bottle's default web server, but not when using mod_wsgi or
Apache.  The problem is that by default, Apache will automatically
decode %2F in URLs *before* your application, and throw 404's.  This
is a known problem for IIIF users and Apache; c.f. the discussion at
https://groups.google.com/g/iiif-discuss/c/VB68Ix4toSM

You can change this behavior by using the Apache configuration flag
`AllowEncodedSlashes`, but doing that is impossible with
mod_wsgi-express, which generates an Apache configuration file and
doens't give you a way to set that particular option.  So, we could
make it work for Bottle in debug mode, and we could tell people who
install DIBS in Apache to set that variable, but we _can't_ make it
work in mod_wsgi-express (which is what `run-server -m test` uses).

But all is not lost.  Since we are reading and rewriting the manifests
already, we can simply translate all instances of the pattern
"barcode%2Fpage" to something that doesn't use `%2F`, such as
"barcode!page", and then when DIBS gets requests on `/iiif/something`,
we can translate any occurrence of "barcode!page" back to
"barcode%2Fpage".  That is the solution implemented here.