Remove tests on __debug__ b/c they can't change things
There are no good situations for optimizing out the calls to log(...),
so there doesn't seem to be a good reason to clutter the code with
tests against __debug__.  It's good for other kinds of software but
DIBS is a case where I think we are unlikely to ever take advantage of
this small potential optimization.  We may as well make the code more
readable instead.