Add additional info to data returned by /item-status
Previously, the availability of an item to the user was based on a
true/false variable, but this referred to whether the user can loan it
out.  In order to check if the item is _currently_ on loan to the
user, it would have been necessary to do a string comparison against
the explanation.  That would have been problematic.  I changed
`/item-status` to return a new field, one that directly indicates
whether the user currently has the item on loan.