Refactor code to use Bottle plugins and fix issue #36
In order to fix issue #36 in (IMHO) the cleanest fashion, it was best
to use Bottle plugins.  In the process of debugging the code and
interactions with normal decorators we used on DIBS routes, I ended up
converting all the decoractors to Bottle plugins.  The new code is
shorter and cleaner, although it does use an advanced Bottle feature
and may be harder to understand at first.  Hopefully this will be okay
in the long run.