AddStaffPersonArgument -> VerifyStaffUser
The AddStaffPersonArgument class was never used in a case where a
person argument needed to be added to a route, which means the name
was confusing, and its goal was poorly thought out.  This commit
changes the name to VerifyStaffUser and change its functionality to be
just to verify that the user has staff priviledges.