Change handling of thumbnail images
The previous approach stored a URL to the thumbnail image for a book.
Sometimes this URL wouldn't work at run-time, and users wouldn't see
the image in the item page.  To avoid this problem, we decided to have
DIBS store the thumbnail images itself.  The new scheme saves the
thumbnails in a separate directory configured using the settings.ini
variable THUMBNAILS_DIR.  The Item object in database.py no longer
needs to have a field for the thumbnail URL.

Folio/Okapi can't supply a URL for a thumbnail image because it's
something that comes from EDS, and EDS doesn't seem to have an API to
retrieve that.  So, I added a new module, thumbnail.py, to deal with
searching for thumbnail images.  The LSP interface layer in lsp.py
handles the requirements of fetching and saving thumbnail images in
whatever way is appropriate for a given LSP.