Avoid scattering config defaults among .py files
Originally, the idea of putting default values for config variables in
places where the config variables are used was as a safety/fall-back
measure, in case something is wrong with settings.ini or values are
not set.  But I'm increasingly of the mind that this is a bad idea,
and that modules should fail with errors if they can't read the
variables they need.