Avoid mixed return types
It's an antipattern to return multiple value types in a case like
human_datetime(), which would normally return a string. It should
really throw an exception if the given value is None or an empty
string, but in the contexts where it's used, it would require the
caller to test, and that would complicate things for little gain in
this case. So, as a compromise, this now returns an empty string
instead of None when the input value is empty.