Manage database connections explicitly
Turns out using the Peewee SqliteDatabase autoconnect feature is
problematic because it leads to the server staying connected to the
database all the time, preventing other processes from writing the
database.  The solution is to connect to the database as needed.  In
the server code, this turns out to be easy to accomplish by defining a
Bottle plugin that wraps every route with a connection.