Change field names in LSPRecord
The fields `id` and `url` are now `item_id` and `item_page`,
respectively.  `id` was probably always a bad idea because it shadows
a Python built-in, and `url` led to easy confusion when reading the
classes in the rest of this file. These names also match the
corresponding field names in the database objects defined in
`data_models.py`, which has benefits for writing code that manipulates
both.